Add config to map topic names to table names
Summary:
The "topics.to.tables" config allows mapping topic names to table names.
The value must be a comma separated list of key value pairs that are
delimited by equal signs. If a topic is not in the map, then the table
name will default to using the table name format.

Test Plan:
```
mvn clean test
```

Reviewers: nn2242, joseph.napolitano, matt.hauser, sam.chase, jason.jho

Reviewed By: jason.jho

Differential Revision: https://phabricator.blueapron.com/D396